+++ miniprogram/setting/setting.js
@@ -3,10 +3,10 @@ module.exports = {
 	CLOUD_ID: 'yunyukeji-d4g7waei5d5d6cdeb', //云服务id ,本地测试环境
 
 	// #### 版本信息
-	VER: 'v1.92',
-	APP_VERSION: '1.92',
+	VER: 'v1.93',
+	APP_VERSION: '1.93',
 	APP_VERSION_DATE: '2026-06-24',
-	APP_VERSION_NAME: '小猫Agent能力目录',
+	APP_VERSION_NAME: 'AI审计统计摘要',
 	COMPANY: '云屿摄影｜技术由云屿科技支持',
 
 	// #### 系统参数